refactor(dynamodb): extract free helper fns from service/mod.rs
service/mod.rs grew to 2931 lines, mostly freestanding helper fns:
PartiQL parsing/evaluation, condition/key-condition/filter expression
evaluators, update-expression apply, attribute-value plumbing, table
description JSON builders. Move all helpers post-AwsService trait impl
to service/helpers.rs and re-export at mod via pub(crate) use.

mod.rs: 2931 -> 424 lines (-86%).
helpers.rs: 2527 lines (pure functions).

Pure refactor, no behavior change. 219 lib tests still pass.
